Add data sources to use SQL queries in ListViews
Using Database::itemDataSource(), one can now create a data source object that
implements the IListDataSource interface. It can be used to conveniently display
the results of an SQLite query in a WinJS.UI.ListView (or FlipView).

Database::groupDataSource() creates an additional data source that can be used
to group the items in the ListView. This can be a bit tricky, because the SQL
queries used in both list views must be compatible -- most importantly, the
items must be ordered exactly as they appear in groups, and they must have a
groupKey property (that can be obtained from a column using the
groupKeyColumnName parameter).