Update README
- Updated SQLite3JS example to use async API
- Stripped down SQLite3 component description and removed the example

The JavaScript API is what I think most people will use this project for, so
it should be featured prominently in the README. We do not have any experience
with using the component in C#, so we just mention it shortly. Actually, there
might be better options for C#/.NET, e.g. sqlite-net.