Enhance transaction support, produces IE script engine error
using this version, a strange problem appears: JSLint validates the code
just fine but WWAHost throws either "SCRIPT1010: Expected identifier", or
sometimes opens the Visual Studio debugger prompt for WWAHost.
The latter happened in SQLite3.js:291.
Another time, a "normal" error in userspace occurred where it complained
about not having started a transaction when trying to commit.

So the code is not complete but this commit should document a problem with
the script engine/WWAHost which would be interesting to reproduce (or not)
on another system.